+++ app/models/competition.rb
@@ -4,6 +4,7 @@ class Competition < ApplicationRecord
   ## relations
   has_many :category_results, dependent: :destroy
   has_many :scores, dependent: :destroy
+  has_many :performed_segments, dependent: :destroy
 
   ## validations
   validates :country, allow_nil: true, format: { with: /\A[A-Z][A-Z][A-Z]\Z/}