+++ app/models/performed_segment.rb
@@ -1,0 +1,6 @@
+class PerformedSegment < ApplicationRecord
+  has_many :scores
+  validates  :starting_time, presence: true
+
+  belongs_to :competition
+end