+++ app/models/score.rb
@@ -8,9 +8,10 @@ class Score < ApplicationRecord
   belongs_to :competition
   belongs_to :skater
   belongs_to :category_result, required: false
+  belongs_to :performed_segment, required: false
 
   ## validations
-  validates  :segment_starting_time, presence: true
+  #validates  :segment_starting_time, presence: true
 
   ## virtual attributes
   def competition_name
@@ -31,6 +32,10 @@ class Score < ApplicationRecord
   def nation
     skater.nation
   end
+  def segment_starting_time
+    performed_segment.starting_time.in_time_zone(competition.timezone)
+  end
+
 =begin
   def elements_summary
     elements.map(&:name).join('/')