+++ db/migrate/002_create_competitions.rb
@@ -33,5 +33,16 @@ class CreateCompetitions < ActiveRecord::Migration[5.1]
       t.belongs_to :competition
       t.references :skater
     end
+
+    ################
+    # performed segments
+    create_table :performed_segments do |t|
+      t.string :category
+      t.string :segment
+      t.datetime :starting_time, default: Time.new(1970, 1, 1, 0, 0, 0)
+      
+      ## relations
+      t.belongs_to :competition
+    end
   end
 end