+++ db/migrate/003_create_scores.rb
@@ -8,7 +8,7 @@ class CreateScores < ActiveRecord::Migration[5.1]
       t.string :category
       t.string :segment
       #t.date :date, default: Date.new(1970, 1, 1)
-      t.datetime :segment_starting_time, default: Time.new(1970, 1, 1, 0, 0, 0)
+      #t.datetime :segment_starting_time, default: Time.new(1970, 1, 1, 0, 0, 0)
       t.string :result_pdf
       
       t.float :tss, default: 0.0
@@ -24,6 +24,7 @@ class CreateScores < ActiveRecord::Migration[5.1]
       t.belongs_to :competition
       t.references :skater
       t.references :category_result
+      t.references :performed_segment
     end
 
     create_table :elements do |t|